+++ simulation/g4simulation/g4eval/EvalLinksV1LinkDef.h
@@ -1,8 +1,8 @@
 #ifdef __CINT__
 
-#pragma link C++ class std::pair<unsigned int,unsigned int>+;
-#pragma link C++ class std::map< std::pair<unsigned int,unsigned int> , float >+;
-#pragma link C++ class std::pair< std::pair<unsigned int,unsigned int> , float >+;
+#pragma link C++ class std::pair<unsigned long long,unsigned long long>+;
+#pragma link C++ class std::map< std::pair<unsigned long long,unsigned long long> , double >+;
+#pragma link C++ class std::pair< std::pair<unsigned long long,unsigned long long> , double >+;
 #pragma link C++ class EvalLinksV1+;
 
 #endif /* __CINT__ */
